EPMEDU-728 - renamed Spacers, moved shortButton to button package
