+++ feature/more/src/main/java/com/epmedu/animeal/more/profile/ProfileScreenUI.kt
@@ -11,7 +11,7 @@ import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import com.epmedu.animeal.base.theme.AnimealTheme
-import com.epmedu.animeal.foundation.topbar.AnimealShortButton
+import com.epmedu.animeal.foundation.button.AnimealShortButton
 import com.epmedu.animeal.foundation.topbar.BackButton
 import com.epmedu.animeal.foundation.topbar.TopBar
 import com.epmedu.animeal.resources.R
